cambios y creacion de tablas en db

+++ src/modules/base/model.py
@@ -7,7 +7,7 @@ from abc import ABC, abstractmethod
 # primary_key=True,
 
 class BaseModel(SQLModel, table=False):
-    id:Optional[int] = Field(sa_column=Column(BigInteger(), default=None, primary_key=True))
+    id:Optional[int] = Field(sa_column=Column(Integer(), default=None, primary_key=True))
     deleted_at:Optional[datetime] = None
     created_at:Optional[datetime] = Field(default_factory=datetime.now,nullable=False)
     updated_at:Optional[datetime] = None
